+++ src/validation/validationWinningNumbers.js
@@ -1,0 +1,31 @@
+import { LOTTO_RULE } from "../constants/lotto.js";
+import { ERROR_MESSAGE } from "../constants/message.js";
+import CustomError from "../CustomError.js";
+import { isInRange, isDuplicate, hasNotInteger } from "../utils/predicate.js";
+
+export const validationWinningNumbers = (winningNumbers) => {
+  if (winningNumbers.length !== LOTTO_RULE.LOTTO_LENGTH) {
+    throw new CustomError(ERROR_MESSAGE.INVALID_LOTTO_LENGTH);
+  }
+
+  if (isDuplicate(winningNumbers)) {
+    throw new CustomError(ERROR_MESSAGE.INVALID_DUPLICATE_NUMBER);
+  }
+
+  if (hasNotInteger(winningNumbers)) {
+    throw new CustomError(ERROR_MESSAGE.INVALID_INTEGER);
+  }
+
+  if (
+    winningNumbers.some(
+      (number) =>
+        !isInRange(
+          number,
+          LOTTO_RULE.MIN_LOTTO_NUMBER,
+          LOTTO_RULE.MAX_LOTTO_NUMBER
+        )
+    )
+  ) {
+    throw new CustomError(ERROR_MESSAGE.INVALID_LOTTO_NUMBER_RANGE);
+  }
+};
